refactor: always load config first, move timescale logic
